// web/src/lib/api-types.ts
export interface components {
  schemas: {
// ...
    ProjectOut: {
      /** Slug */
      slug: string;
      /** Title */
      title: string;
      /** Description */
      description: string;
      /** Content */
      content?: string | null;
      /** Tech Stack */
      tech_stack?: string[];
      /** Image Url */
      image_url?: string | null;
      /** Repo Url */
      repo_url?: string | null;
      /** Live Url */
      live_url?: string | null;
      /**
       * Display Order
       * @default 0
       */
      display_order: number;
      /**
       * Is Published
       * @default false
       */
      is_published: boolean;
      /** Id */
      id: number;
      /**
       * Created At
       * Format: date-time
       */
      created_at: string;
      /**
       * Updated At
       * Format: date-time
       */
      updated_at: string;
    };
// ...
  };
  responses: never;
  parameters: never;
  requestBodies: never;
  headers: never;
  pathItems: never;
}